Allow teams without members
On the accounts team, we have two slack channels:

* govuk-account-tech - a channel specifically for the developers.
* govuk-accounts - the channel for the whole multidisciplinary team.

We want PRs to alert in govuk-accounts-tech, but team quotes to alert in
govuk-accounts. The govuk-accounts team is configured without team
members, but with quotes and a slack channel.

Current logic short-circuits if a team has no team members. The intent
of this was to stop the seal alerting with every PR when someone pings
the seal from a slack channel that isn't defined in the config.

Changing the logic to test for a `slack_channel` instead of `members`
should retain the short circuit but also allow us to have teams without
team members.