Remove GOV.UK Design System team :sob:
In a retro we found that most our team never look at the message from
the sealbot :disappointed_relieved:

Since we're not finding it useful, and there's a cost associated with
keeping the team updated on here, we're removing ourselves from the
configuration.

So long and thanks for all the ~fish~ PRs! :wave: :seal: