Get repos from JSON data if they aren't defined
If a team does not contain a list of repos in the static config,
they will be queried for in the data returned from the repos.json
file.

There are also some instances where the team Slack channel differs
to the team name e.g. `govuk-platform-reliability-team` vs.
`govuk-platform-reliability` so we need use the channel name.