Update Readme
The "Config" section should hopefully be clearer now about whether to
add the repos to the config file depending on whether it's part of the
GOV.UK programme or not.

The "Local testing" section is out of date and most people don't have
Slack admin permissions to change the webhook settings.