Add reminder about reward nominations
It's all too easy to forget this is a thing, and/or not know where to find more information. Putting it in a Slack reminder should keep it on peoples' radar.

Opted to post on Wednesdays as a number of people have non-working days on Fridays, and messages sent on Mondays may get lost amongst all the other noise at the beginning of the week.