Add a new reporter for all bots
Aim here is that all bot PRs get reported for someone to review and label/action them.

Particular use-case is to add maintenance/security upgrade labels to dependabot PRs and those labels will be used to slim down what is reported as a security thing for first responders.  But we're adding the sweep bot too so we can see the PRs it's generating and decide what to do with them (they otherwise seem quite hidden).