Make github_bots team be barked at
Put this one first.  Barking is slow and so there's a slim chance that by the time we bark at that team, someone _could_ review the PRs and flag them as maintenance-upgrade before the security-alerts team is barked at which would then exclude the PR from the report.  Admittedly a slim chance, but possible.