Add Pricing & Lending devs (#10)
Add configuration for Pricing & Lending devs. Yes, the naming right now
is a bit confusing: `pricing_and_lending`, `ds-payments-and-lending`. We
are changing names with the change of term, some bits have already been
renamed, but we are not 100% sure of what the final name will be, so we
are holding off renaming Github team names.