Don't rescue MessageBuilder errors
The error handling that previously rescued MessageBuilder errors in the GitHub
Actions workflow resulted in 30-35 error notifications in the
didn't catch errors with the Slack integration.

It's no longer needed since we introduced a new job to alert on failures in the
workflows.

The only potential downside is that the script will stop if a single message
fails, but based on past experience this is a very unlikely scenario that the
issue is confined to a single message.