Ignore search-api-v2-dataform
The search team are getting the following error every morning:
1 errors fetching security alerts

It looks as though this is due to the fact that [search-api-v2-dataform]
is listed as a repo they own, but does not have dependabot set up.
Dependabot is not needed because it is a repo of sql queries and does
not have any dependencies.

[search-api-v2-dataform]: https://github.com/alphagov/govuk-developer-docs/blob/main/data/repos.yml#L695